+++ bibliotheque - 3/src/ca/qc/collegeahuntsic/bibliotheque/dao/LivreDAO.java
@@ -71,7 +71,7 @@ public class LivreDAO extends DAO {
     public void add(LivreDTO livreDTO) throws DAOException {
         try(
             PreparedStatement addPreparedStatement = getConnection().prepareStatement(LivreDAO.ADD_REQUEST)) {
-            addPreparedStatement.setInt(1,
+            addPreparedStatement.setLong(1,
                 getPrimaryKey());
             addPreparedStatement.setString(2,
                 livreDTO.getTitre());
@@ -253,7 +253,7 @@ public class LivreDAO extends DAO {
      * @return La clé primaire généré
      * @throws DAOException S'il y a une erreur avec la base de données
      * */
-    private int getPrimaryKey() throws DAOException {
+    private long getPrimaryKey() throws DAOException {
         return getPrimaryKey(LivreDAO.CREATE_PRIMARY_KEY);
     }
 }
