+++ bibliotheque - 3/src/ca/qc/collegeahuntsic/bibliotheque/dao/PretDAO.java
@@ -339,7 +339,7 @@ public class PretDAO extends DAO {
      * @return La clé primaire généré
      * @throws DAOException S'il y a une erreur avec la base de données
      * */
-    private int getPrimaryKey() throws DAOException {
+    private long getPrimaryKey() throws DAOException {
         return getPrimaryKey(PretDAO.CREATE_PRIMARY_KEY);
     }
 }
